fastapi_app/requirements.txt
fastapi
pydantic
uvicorn
torch
torchvision
Pillow